Move default channel menu header next to ChannelActionsSheet
Moves DefaultChannelMenuHeaderContent (renamed from
DefaultSelectedChannelMenuHeaderContent) and its private helpers to
ChannelActionsSheet.kt. When SelectedChannelMenu is removed in v8, the
old file becomes a single git rm.

No behavior change — the factory still delegates to the same function.
