+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/channels/info/ChannelActionsSheet.kt
@@ -16,15 +16,40 @@
 
 package io.getstream.chat.android.compose.ui.channels.info
 
+import androidx.annotation.DrawableRes
+import androidx.annotation.StringRes
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import io.getstream.chat.android.client.extensions.isMutedFor
+import io.getstream.chat.android.client.extensions.isPinned
+import io.getstream.chat.android.compose.R
 import io.getstream.chat.android.compose.ui.components.StreamCardBottomSheet
+import io.getstream.chat.android.compose.ui.components.avatar.AvatarSize
+import io.getstream.chat.android.compose.ui.theme.ChannelAvatarParams
 import io.getstream.chat.android.compose.ui.theme.ChannelMenuCenterContentParams
 import io.getstream.chat.android.compose.ui.theme.ChannelMenuHeaderContentParams
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
+import io.getstream.chat.android.compose.ui.theme.StreamTokens
+import io.getstream.chat.android.compose.ui.util.dmCounterpartId
+import io.getstream.chat.android.compose.ui.util.getMembersStatusText
 import io.getstream.chat.android.models.Channel
 import io.getstream.chat.android.models.User
 import io.getstream.chat.android.previewdata.PreviewChannelData
@@ -83,6 +108,130 @@ public fun ChannelActionsSheet(
         header()
         content()
     }
+}
+
+/**
+ * Default header for the channel actions menu. Wired in by the
+ * `ChatComponentFactory.ChannelMenuHeaderContent` factory method and shared by
+ * [ChannelActionsSheet] and the deprecated [SelectedChannelMenu].
+ *
+ * Renders inline muted and pinned icons next to the channel name based on the channel's pin state
+ * and the current user's mute settings. When [currentUser] is `null`, no state icons are rendered.
+ *
+ * @param selectedChannel The channel the user selected.
+ * @param currentUser The currently logged-in user data.
+ * @param modifier Modifier applied to the header row. The deprecated SelectedChannelMenu uses
+ * this to add a top inset (its Card has no drag handle); ChannelActionsSheet leaves it empty
+ * since the Material 3 drag handle already provides the top spacing.
+ */
+@Composable
+internal fun DefaultChannelMenuHeaderContent(
+    selectedChannel: Channel,
+    currentUser: User?,
+    modifier: Modifier = Modifier,
+) {
+    val showPinnedIcon = selectedChannel.isPinned()
+    val showMutedIcon = currentUser != null && isChannelOrCounterpartMuted(selectedChannel, currentUser)
+    Row(
+        modifier = modifier
+            .fillMaxWidth()
+            .padding(
+                start = StreamTokens.spacingMd,
+                end = StreamTokens.spacingMd,
+                bottom = StreamTokens.spacingSm,
+            ),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        ChatTheme.componentFactory.ChannelAvatar(
+            params = ChannelAvatarParams(
+                modifier = Modifier.size(AvatarSize.ExtraLarge),
+                channel = selectedChannel,
+                currentUser = currentUser,
+                showIndicator = true,
+            ),
+        )
+
+        Column(
+            modifier = Modifier
+                .padding(start = StreamTokens.spacingSm)
+                .weight(1f),
+        ) {
+            HeaderTitleRow(
+                selectedChannel = selectedChannel,
+                currentUser = currentUser,
+                showMutedIcon = showMutedIcon,
+                showPinnedIcon = showPinnedIcon,
+            )
+            Text(
+                text = selectedChannel.getMembersStatusText(
+                    context = LocalContext.current,
+                    currentUser = currentUser,
+                ),
+                style = ChatTheme.typography.captionDefault,
+                color = ChatTheme.colors.textSecondary,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+            )
+        }
+    }
+}
+
+private fun isChannelOrCounterpartMuted(channel: Channel, currentUser: User): Boolean {
+    if (channel.isMutedFor(currentUser)) return true
+    val otherUserId = channel.dmCounterpartId(currentUser) ?: return false
+    return currentUser.mutes.any { it.target?.id == otherUserId }
+}
+
+@Composable
+private fun HeaderTitleRow(
+    selectedChannel: Channel,
+    currentUser: User?,
+    showMutedIcon: Boolean,
+    showPinnedIcon: Boolean,
+) {
+    Row(
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.spacedBy(StreamTokens.spacing2xs),
+    ) {
+        Text(
+            modifier = Modifier.weight(1f, fill = false),
+            text = ChatTheme.channelNameFormatter.formatChannelName(selectedChannel, currentUser),
+            style = ChatTheme.typography.headingSmall,
+            color = ChatTheme.colors.textPrimary,
+            maxLines = 1,
+            overflow = TextOverflow.Ellipsis,
+        )
+        if (showMutedIcon) {
+            HeaderStateIcon(
+                iconRes = R.drawable.stream_design_ic_mute,
+                contentDescriptionRes = R.string.stream_compose_channel_item_muted,
+                testTag = "Stream_ChannelMenuHeaderMutedIcon",
+            )
+        }
+        if (showPinnedIcon) {
+            HeaderStateIcon(
+                iconRes = R.drawable.stream_design_ic_pin,
+                contentDescriptionRes = R.string.stream_compose_channel_item_pinned,
+                testTag = "Stream_ChannelMenuHeaderPinnedIcon",
+            )
+        }
+    }
+}
+
+@Composable
+private fun HeaderStateIcon(
+    @DrawableRes iconRes: Int,
+    @StringRes contentDescriptionRes: Int,
+    testTag: String,
+) {
+    Icon(
+        modifier = Modifier
+            .testTag(testTag)
+            .size(16.dp),
+        painter = painterResource(id = iconRes),
+        contentDescription = stringResource(contentDescriptionRes),
+        tint = ChatTheme.colors.textTertiary,
+    )
 }
 
 @Preview(showBackground = true)
