+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/theme/ChatComponentFactory.kt
@@ -64,7 +64,7 @@ import io.getstream.chat.android.compose.ui.attachments.content.onFileAttachment
 import io.getstream.chat.android.compose.ui.channel.info.ChannelInfoNavigationIcon
 import io.getstream.chat.android.compose.ui.channels.header.DefaultChannelListHeaderTrailingContent
 import io.getstream.chat.android.compose.ui.channels.info.ChannelActionsSheet
-import io.getstream.chat.android.compose.ui.channels.info.DefaultSelectedChannelMenuHeaderContent
+import io.getstream.chat.android.compose.ui.channels.info.DefaultChannelMenuHeaderContent
 import io.getstream.chat.android.compose.ui.channels.list.ChannelItem
 import io.getstream.chat.android.compose.ui.channels.list.DefaultChannelItemCenterContent
 import io.getstream.chat.android.compose.ui.channels.list.DefaultChannelItemLeadingContent
@@ -1815,7 +1815,7 @@ public interface ChatComponentFactory {
      */
     @Composable
     public fun ChannelMenuHeaderContent(params: ChannelMenuHeaderContentParams) {
-        DefaultSelectedChannelMenuHeaderContent(
+        DefaultChannelMenuHeaderContent(
             selectedChannel = params.selectedChannel,
             currentUser = params.currentUser,
             modifier = params.modifier,
